src/main/java/com/example/demo/validation/PhoneNumberValidator.java
package com.example.demo.validation;


import jakarta.validation.ConstraintValidator;
import jakarta.validation.ConstraintValidatorContext;

public class PhoneNumberValidator implements ConstraintValidator<ValidPhoneNumber, String> {

    @Override
    public void initialize(ValidPhoneNumber constraintAnnotation) {
    }

    @Override
    public boolean isValid(String phoneNumber, ConstraintValidatorContext context) {
        if (phoneNumber == null) {
            return false;
        }

        String regex = "^[1-9][0-9]{0,2}[0-9]{7,14}$";

        return phoneNumber.matches(regex);
    }
}